// Application.RMGUI/ClientApp/src/app/models/ProjectAttribute.ts
export class ProjectAttribute {

  project_Name: string;
  project_Code: string;
  project_Description: string;
  project_StartDate: string;
  project_EndDate: string;
  project_Status: string;
  project_LocationId: string;
  project_Location: string;
  project_Billable: string;
  customer_Details: string;
  created_By: string;
  created_On: string;
}